+++ src/main/webapp/components/login/index.js
@@ -1,0 +1,4 @@
+import LoginGQL from './container'
+import Login from './login'
+
+export { LoginGQL as default, Login }
