+++ src/main/webapp/components/login/loginModal.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Dialog from '@material-ui/core/Dialog'
 import Container from './container'
 
-export const LogInModal = props => {
+const LogInModal = props => {
   const { handleClose, open, label } = props
   return (
     <Dialog open={open}>
